Move testing into docker for better consistency.
