+++ pythonx/UltiSnips/snippet/definition/__init__.py
@@ -1,4 +1,4 @@
 """In memory representation of snippet definitions."""
 
-from UltiSnips.snippet.definition.ultisnips import UltiSnipsSnippetDefinition
+from UltiSnips.snippet.definition.ulti_snips import UltiSnipsSnippetDefinition
 from UltiSnips.snippet.definition.snipmate import SnipMateSnippetDefinition
