+++ pythonx/UltiSnips/snippet/definition/ulti_snips.py
@@ -4,7 +4,7 @@
 """A UltiSnips snippet after parsing."""
 
 from UltiSnips.snippet.definition._base import SnippetDefinition
-from UltiSnips.snippet.parsing.ultisnips import parse_and_instantiate
+from UltiSnips.snippet.parsing.ulti_snips import parse_and_instantiate
 
 
 class UltiSnipsSnippetDefinition(SnippetDefinition):
