+++ pythonx/UltiSnips/snippet/source/__init__.py
@@ -6,5 +6,5 @@
 from UltiSnips.snippet.source._base import SnippetSource
 from UltiSnips.snippet.source.added import AddedSnippetsSource
 from UltiSnips.snippet.source.file.snipmate import SnipMateFileSource
-from UltiSnips.snippet.source.file.ultisnips import UltiSnipsFileSource, \
+from UltiSnips.snippet.source.file.ulti_snips import UltiSnipsFileSource, \
     find_all_snippet_files, find_snippet_files
