+++ scripts/install_packages.sh
@@ -1,0 +1,11 @@
+#!/bin/sh
+
+set -o errexit
+set -o verbose
+
+apt-get update
+apt-get install -y \
+    g++ \
+    tmux \
+    git
+apt-get clean
