+++ test/vim_test_case.py
@@ -83,9 +83,8 @@ class VimTestCase(unittest.TestCase, TempFileManager):
         os.symlink(source, os.path.join(absdir, os.path.basename(source)))
 
     def setUp(self):
-        # TODO(sirver): this uses 'vim', but must use --vim from the commandline.
         if not VimTestCase.version:
-            VimTestCase.version, _ = subprocess.Popen(['vim', '--version'],
+            VimTestCase.version, _ = subprocess.Popen([self.vim.vim_executable, '--version'],
                                                       stdout=subprocess.PIPE, stderr=subprocess.PIPE).communicate()
             if PYTHON3:
                 VimTestCase.version = VimTestCase.version.decode('utf-8')
